Big changes to the backend and the frontend.

+++ Core/Interfaces/ISearchTypeImplementationFactory.cs
@@ -4,5 +4,5 @@ namespace Core.Interfaces;
 
 public interface ISearchTypeImplementationFactory
 {
-    ISimilarFilesFinder GetSearchImplementation(FileSearchType searchType, double degreeOfSimilarity = 0);
+    ISimilarFilesFinder GetSearchImplementation(FileSearchType searchType, int degreeOfSimilarity = 0);
 }
